langchain/pyproject.toml
[tool.poetry]
name = "langchain_projects"
version = "0.1.0"
description = ""
authors = ["pablo-lucas <pablo.lucv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.324"
streamlit = "^1.28.0"
openai = "^0.28.1"
tiktoken = "^0.5.1"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.4"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
